requirements.txt
python-nmap
mysql-connector-python
flask